potential fix for LP #1879975
In compiz::private_screen::EventManager::removeWatchFd(),
the CompWatchFd objects are only removed from the internal list, but are not
detached from the Glib::MainContext that they were attached to previously. I
believe this results in the Glib event loop still calling the original
callbacks registered on them.

This fixes the bug for me.

It has some regression potential if there is a component that relies on this
behavior. In my opinion this is not likely given that the CompWatchFd objects
are only created in EventManager::addWatchFd() and not reference to them is
kept after they are removed.