dbus: need to use dbus_bool_t instead of bool
Passing the address of a bool variable to a C function is undefined behavior.
Notably, the dbus library expects a dbus_bool_t, which is an alias for uint32_t,
but checks that the value is either zero or one. On the other hand, the C++
compiler is free to use a smaller size object to represent a bool and also
free to use any value for true.